03-emailAutomation/requirements.txt
black
mypy
python-dotenv